% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{dostuff}
\alias{dostuff}
\title{A short description.}
\usage{
dostuff(x)
}
\arguments{
\item{x}{Description of the parameter.}
}
\value{
The returned value.
}
\description{
More detailed description
}
\examples{
x <- rnorm(1000)
}
\seealso{
\code{\link{bla}} for doing it slower.
}
\keyword{basis}
\keyword{k-cone}

