% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_info.R
\docType{data}
\name{files}
\alias{files}
\title{File infos for the GDC release v2 from 08-2016}
\format{A data frame with the following columns.
\describe{
  \item{data_type}{The type of the file.}
  \item{updated_datetime}{Last update of the file.}
  \item{file_name}{The filename of the file.}
  \item{md5sum}{md5 hash for the file.}
  \item{file_id}{The GDC UUID for the file.}
  \item{data_category}{Category for the file.}
  \item{experimental_strategy}{Type of experiment used to obtain the data.}
  \item{patient_uuid}{Unique ID for the patient the data came from.}
  \item{barcode}{Barcode for the patient the data came from.}
}}
\usage{
files
}
\description{
This data set contains detailed information about all files in the GDC
release that have a corresponding read method in `tcgar`
}
\keyword{datasets}

