% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rnaseq.R
\docType{data}
\name{genemap}
\alias{genemap}
\title{Gene annotations obatined from BioMart.}
\format{A data frame with 66,203 rows and 4 variables:
\describe{
  \item{ensgene}{Ensembl Gene ID}
  \item{description}{Description of the Gene}
  \item{symbol}{The Gene symbol}
  \item{entrez}{Entrez Gene ID}
}}
\usage{
genemap
}
\description{
This data set contains additional annotations for genes
obtained from BioMart.
}
\keyword{datasets}

