% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Downloads data from the Genomic Data Commons Portal.}
\usage{
get_data(manifest, out = "gdc_data", quiet = FALSE, np = 4)
}
\arguments{
\item{manifest}{Path to the GDC file manifest.}

\item{out}{Optional. The folder to which to download the data. Default
is to extract into a folder named "gdc_data".}

\item{quiet}{Specifies whether download status info should be supressed (default no).}

\item{np}{How many parallel download processes should be started.}
}
\value{
Nothing.
}
\description{
This will download all data specified in a given manifest file. To generate
a valid manifest use the portal at \url{https://gdc-portal.nci.nih.gov/}.
}
\examples{
# Not run due to large download...
# gbm <- system.file("extdata", "manifest.tsv", package = "tcgar")
# d <- tempdir()
# get_data(gbm, out=d)

}
\keyword{GDC}
\keyword{TCGA,}
\keyword{download,}
