% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_panels.R
\name{get_panels}
\alias{get_panels}
\title{Lists all available cancer panels in the public TCGA repository.}
\usage{
get_panels()
}
\value{
A character vector containing the available cancer panels.
}
\description{
\emph{Note that you should not extract several downloads of the same technology 
to the same folder as this will be incompatible with the \code{read_*} 
functions.}
}
\examples{
gbm <- system.file("extdata", "GBM", package = "tcgar")
rna <- read_rnaseq(gbm)

}
\seealso{
\code{\link{get}} to download data for a selected cancer panel.
}

