% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_huex.R
\docType{data}
\name{huex_bm}
\alias{huex_bm}
\title{Gene annotations for the TCGA HuEx data.}
\format{A data frame with 18632 rows and 4 variables:
\describe{
  \item{ensgene}{Ensembl Gene ID}
  \item{description}{Description of the Gene}
  \item{symbol}{The Gene symbol}
  \item{entrez}{Entrez Gene ID}
}}
\usage{
huex_bm
}
\description{
This data set contains additional annotations for the HuEx features obtained
from BioMart.
}
\keyword{datasets}

