% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_stats.R
\name{panel_samples}
\alias{panel_samples}
\title{Counts samples per panel across several data sets.}
\usage{
panel_samples(...)
}
\arguments{
\item{...}{Several or one data set as returned by the `read_*` functions.}
}
\value{
A list of length(what) containing the combined data or NULL if no data
 was found.
}
\description{
Counts samples per panel across several data sets.
}
\examples{
gbm <- system.file("extdata", "manifest.tsv", package = "tcgar")
d <- tempdir()
panel_samples(huex=read_huex(gbm, d, progress=FALSE))

}
\keyword{TCGA}
\keyword{statistics}

