% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mut_dist.R
\name{plot_mut_dist}
\alias{plot_mut_dist}
\title{Plot mutations distributions}
\usage{
plot_mut_dist(gene_df)
}
\arguments{
\item{gene_df}{data.frame}
}
\value{
ggplot
}
\description{
Plot mutations distributions
}
\examples{

}

