% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mut_freq.R
\name{plot_mut_freq}
\alias{plot_mut_freq}
\title{Plot mutation frequencies}
\usage{
plot_mut_freq(gene_df)
}
\arguments{
\item{gene_df}{data.frame}
}
\value{
ggplot
}
\description{
Plot mutation frequencies
}
\examples{
plot_mut_freq(read_cbioportal("~/mutation_table_TP53.csv"))
}

