% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bulk.R
\name{read_bulk}
\alias{read_bulk}
\title{Reads several data types from several TCGA download directories.}
\usage{
read_bulk(folders, what = c("HuEx", "RNASeqV2", "clinical"), args = list(),
  stats = TRUE)
}
\arguments{
\item{folders}{A character vector of folders containing downloaded TCGA data.}

\item{what}{What technologies to read. Must be a character vector consisting 
of any of the following: "HuEx", "RNASeqV2", "clinical".}

\item{args}{A named list containing additional arguments to the function. Only
the \code{folder} argument will be filled in by the bulk reader.}

\item{stats}{Logical. Whether to print some basic stats about the bulk read.}
}
\value{
A list of length(what) containing the combined data or NULL if no data
 was found.
}
\description{
\code{read_bulk} allows you to combine several data types across several
directories containing TCGA data. Data from several directories will be 
combined correctly automatically.
}
\examples{
gbm <- system.file("extdata", "GBM", package = "tcgar")
rna <- read_bulk(gbm)

importFrom data.table rbindlist
}
\keyword{TCGA}
\keyword{bulk}
\keyword{read}

