% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cbioportal.R
\name{read_cbioportal}
\alias{read_cbioportal}
\title{Read in data from cbioportal}
\usage{
read_cbioportal(gene_file)
}
\arguments{
\item{gene_file}{.tsv}
}
\value{
data.frame
}
\description{
Input must be data downloaded from http://www.cbioportal.org/ by selecting cancer studies
of interest and inputting one gene name. The download button for the required csv is under
the Mutations tab on the bottom left of the resulting table.
}
\examples{

}

