% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clinical.R
\name{read_clinical}
\alias{read_clinical}
\title{Reads clinical data from TCGA for multiple samples.}
\usage{
read_clinical(manifest, folder, additional = NULL, progress = TRUE)
}
\arguments{
\item{manifest}{Path to the GDC file manifest.}

\item{folder}{The folder that contains the data.}

\item{additional}{A named vector with entries `name = xpath` where `name` is
the column name that should be extracted and `xpath` is an XPath expression
defining the XML tag that should be read. Entries will returned as
unmodified strings.}

\item{progress}{Whether to show progress information.}
}
\value{
A data table containing the information for patients on its rows.
}
\description{
This function assembles a data set with the major clinical indicators for each
patient as well as sample procurement data for all the samples available for
each patient from TCGA Biotab data.
}
\examples{
# Not run due to large download...
# gbm <- system.file("extdata", "manifest.tsv", package = "tcgar")
# d <- tempdir()
# clin <- read_clinical(gbm, d)

}
\keyword{GDC}
\keyword{TCGA}
\keyword{clinical}
\keyword{read}
