% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rnaseq.R
\name{read_rnaseq}
\alias{read_rnaseq}
\title{Reads RNA sequencing data from TCGA for several samples.}
\usage{
read_rnaseq(folder, features = "genes", normalization = "raw",
  progress = FALSE)
}
\arguments{
\item{folder}{Folder where the data files reside.}

\item{features}{The feature type. Must be one of "genes", "isoforms",
"junctions" or "exons".}

\item{normalization}{The normalization method. Must be one of
\describe{
    \item{"raw"}{The raw counts.}
    \item{"Q75"}{Normalization by dividing through the 75% quantile of the
    raw counts (TCGA default). This only available for transcripts and
    isoforms.}
    \item{"XPM"}{X per million. This will use transcripts per million (TPM)
    for genes and isoforms and reads per kilobase of transcript per million
    (RPKM) for junctions and exons.}
}}

\item{progress}{Logical. Show progress info?}
}
\value{
A data table containing the features as rows and the samples in the
 columns.
}
\description{
Reads RNASeqV2 Level 3 data downloaded from TCGA. The functions does some
conservative checking in order to validate that all samples and features are
annotated correctly. Also note that duplicate data (files with the same
extract name) are dealt with implicitly. This will for instance occur if you
have data for several Illumina technologies in the same directory (for TCGA
mostly IlluminaGA or IlluminaHiSeq). The tie breaker used is to prefer HiSeq
or use the last of the sorted file names if HiSeq is not available.
}
\examples{
gbm <- system.file("extdata", "GBM", package = "tcgar")
rna <- read_rnaseq(gbm)

}
\seealso{
\code{\link{read_huex}} to read exon expression data.
}
\keyword{RNASeq}
\keyword{TCGA}
\keyword{read}

