% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rnaseq_legacy.R
\name{read_rnaseq_legacy}
\alias{read_rnaseq_legacy}
\title{Reads RNA sequencing data from the GDC legacy archive.}
\usage{
read_rnaseq_legacy(manifest, folder, features = "genes",
  normalization = "raw", progress = TRUE)
}
\arguments{
\item{manifest}{Path to the GDC file manifest.}

\item{folder}{Folder where the data files reside.}

\item{features}{The feature type. Must be one of "genes", "isoforms",
"junctions" or "exons".}

\item{normalization}{The normalization method. Must be one of
\describe{
    \item{"raw"}{The raw counts.}
    \item{"Q75"}{Normalization by dividing through the 75% quantile of the
    raw counts (TCGA default). This only available for transcripts and
    isoforms.}
    \item{"XPM"}{X per million. This will use transcripts per million (TPM)
    for genes and isoforms and reads per kilobase of transcript per million
    (RPKM) for junctions and exons.}
}}

\item{progress}{Logical. Show progress info?}
}
\value{
A data table containing the features as rows and the samples in the
 columns.
}
\description{
Reads RNASeq Level 3 data downloaded from the legacy GDC. The functions does some
conservative checking in order to validate that all samples and features are
annotated correctly. Also note that duplicate data (files with the same
extract name) will be treated as independent duplicates for the same sample.
}
\examples{
# Not run due to large download...
# gbm <- system.file("extdata", "manifest.tsv", package = "tcgar")
# d <- tempdir()
# rna <- read_rnaseq(gbm, d)

}
\seealso{
\code{\link{read_huex}} to read exon expression data.
}
\keyword{GDC}
\keyword{RNASeq}
\keyword{TCGA}
\keyword{read}
