% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rnaseq.R
\docType{data}
\name{rnaseq_bm}
\alias{rnaseq_bm}
\title{Gene annotations for the TCGA RNASeqV2 data.}
\format{A data frame with 20531 rows and 4 variables:
\describe{
  \item{ensgene}{Ensembl Gene ID}
  \item{description}{Description of the Gene}
  \item{symbol}{The Gene symbol}
  \item{entrez}{Entrez Gene ID}
}}
\usage{
rnaseq_bm
}
\description{
This data set contains additional annotations for the RNASeq features
obtained from BioMart.
}
\keyword{datasets}

