\name{tcgaR-package}
\alias{tcgaR-package}
\alias{tcgaR}
\docType{package}
\title{
Interface in R for the TCGA Portal
}
\description{
R interface for the The Cancer Genome Atlas (TCGA) portal to download methylation data from Illumina 27k and 450k arrays. 
}
\author{
Jean-Philippe Fortin <jfortin@jhsph.edu>
}
\keyword{ package }
