% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcgaR-package.r
\docType{package}
\name{tcgaR}
\alias{tcgaR}
\alias{tcgaR-package}
\title{tcgaR:}
\description{
Cancer is hallmarked by many mutations that vary across different types.
TCGA is a large-scale effort in bringing together genomic data on all cancer types;
this package will use mutation data to produce preliminary plots to gain basic insights
into our genes of interest.
}

