% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Downloads data for a given cancer panel and technology.}
\usage{
get_data(panel, tech = "clinical", out = toupper(panel), extract = TRUE,
  method = "auto", quiet = FALSE)
}
\arguments{
\item{panel}{Name of the cancer panel. See \code{\link{get_panels}} to get
a list of availabe cancer panels.}

\item{tech}{The technology for which data should be downloaded. Currently only
supports 'HuEx', 'RNASeqV2' and 'clinical'. Can be a single technology or
a character vector of technologies.}

\item{out}{Optional. The folder to which to extract the data. Default
is to extract into a folder named after the cancer panel within the current
working directory.}

\item{extract}{Boolean. Whether to extract the data after download.}

\item{method}{Passed to \code{\link{download.file}}. Specifies the download
method. Defaults to 'auto'.}

\item{quiet}{Passed to \code{\link{download.file}}. Specifies whether download
status info should be supressed. Defaults to false.}
}
\value{
Nothing.
}
\description{
This will download all data for a given cancer panel and technology as
a given cancer type and extract it.
}
\details{
\emph{Note that you should not extract several downloads of the same technology
to the same folder as this will be incompatible with the \code{read_*}
functions.}
}
\examples{
d <- tempdir()
get_data("GBM", out=d)

}
\seealso{
\code{\link{read_bulk}} to read data from several folders and
 technologies.
}
\keyword{TCGA}
\keyword{download,}

