% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_huex.R
\name{read_huex}
\alias{read_huex}
\title{Reads HuEx 1.0 st v2 exon expression data from the TCGA project.}
\usage{
read_huex(folder, features = "genes")
}
\arguments{
\item{folder}{Location of the directory containing downloaded TCGA data.}

\item{features}{Which features to use. Can be either "genes" or "exons".}
}
\value{
A list with two elements, a matrix of log-expression values and a
 data table giving additional information for samples.
}
\description{
This reads level 3 data for exon expression data for the TCGA project.
The function is currently limited to level 3 data since lower level data
are not publically available.
}
\examples{
gbm <- system.file("extdata", "GBM", package = "tcgar")
huex <- read_huex(gbm)

}
\seealso{
\code{\link{read_rnaseq}} to read RNASeq data.
}
\keyword{HuEx}
\keyword{TCGA}
\keyword{microarray}

