% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_huex.R
\name{read_huex}
\alias{read_huex}
\title{Reads HuEx 1.0 st v2 exon expression data.}
\usage{
read_huex(manifest, folder, features = "genes", progress = TRUE)
}
\arguments{
\item{manifest}{Path to the GDC file manifest.}

\item{folder}{Location of the directory containing downloaded TCGA data.}

\item{features}{Which features to use. Can be either "genes" or "exons".}

\item{progress}{Whether to show progress information.}
}
\value{
A list with three elements, a matrix of log-expression values, a
 data table giving additional information for samples and a data table
 describing the features (genes or exon IDs).
}
\description{
This reads level 3 data for exon expression data.
}
\examples{
# Not run due to large download...
# gbm <- system.file("extdata", "manifest.tsv", package = "tcgar")
# d <- tempdir()
# huex <- read_huex(gbm, d)

}
\seealso{
\code{\link{read_rnaseq}} to read RNASeq data.
}
\keyword{GDC}
\keyword{HuEx}
\keyword{TCGA}
\keyword{microarray}
