% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rnaseq.R
\name{read_rnaseq}
\alias{read_rnaseq}
\title{Reads RNA sequencing data from TCGA for several samples.}
\usage{
read_rnaseq(folder, features = "genes", normalization = "raw")
}
\arguments{
\item{folder}{Folder where the data files reside.}

\item{features}{The feature type. Must be one of "genes", "isoforms",
"junctions" or "exons".}

\item{normalization}{The normalization method. Must be one of
\describe{
    \item{"raw"}{The raw counts.}
    \item{"Q75"}{Normalization by dividing through the 75% quantile of the
    raw counts (TCGA default). This only available for transcripts and 
    isoforms.}
    \item{"XPM"}{X per million. This will use transcripts per million (TPM)
    for genes and isoforms and reads per kilobase of transcript per million
    (RPKM) for junctions and exons.}
}}
}
\value{
A data table containing the features as rows and the samples in the
 columns.
}
\description{
More description...
}
\examples{
NULL

}
\seealso{
\code{\link{read_huex}} to read exon expression data.
}
\keyword{RNASeq}
\keyword{TCGA}
\keyword{read}

